requirements.txt
Django>=5.1,<7.0
djangorestframework>=3.15
drf-spectacular>=0.27
djangorestframework-simplejwt>=5.3
psycopg[binary]>=3.2
